fix: fix type
