+++ packages/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/type.ts
@@ -2,7 +2,10 @@ import type { IBackgroundApi } from '@onekeyhq/kit-bg/src/apis/IBackgroundApi';
 import type { IAirGapUrJson } from '@onekeyhq/qr-wallet-sdk';
 import type { IServerNetwork } from '@onekeyhq/shared/types';
 import type { INetworkAccount } from '@onekeyhq/shared/types/account';
-import type { EQRCodeHandlerType } from '@onekeyhq/shared/types/qrCode';
+import type {
+  EQRCodeHandlerNames,
+  EQRCodeHandlerType,
+} from '@onekeyhq/shared/types/qrCode';
 import type { ITokenData } from '@onekeyhq/shared/types/token';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -104,18 +107,6 @@ export type IQRCodeHandler<T extends IBaseValue> = (
 
 export type IQRCodeHandlerParseResult<T extends IBaseValue> =
   IQRCodeHandlerResult<T> & { raw: string };
-
-export enum EQRCodeHandlerNames {
-  bitcoin = 'bitcoin',
-  ethereum = 'ethereum',
-  solana = 'solana',
-  walletconnect = 'walletconnect',
-  migrate = 'migrate',
-  animation = 'animation',
-  urlAccount = 'urlAccount',
-  marketDetail = 'marketDetail',
-  sendProtection = 'sendProtection',
-}
 
 export type IQRCodeHandlerParseOutsideOptions = {
   handlers: EQRCodeHandlerNames[];
