+++ packages/kit/src/components/HyperlinkText/index.tsx
@@ -4,7 +4,7 @@ import { useMemo } from 'react';
 import { useIntl } from 'react-intl';
 
 import { type ISizableTextProps, SizableText } from '@onekeyhq/components';
-import { EQRCodeHandlerNames } from '@onekeyhq/kit-bg/src/services/ServiceScanQRCode/utils/parseQRCode/type';
+import { EQRCodeHandlerNames } from '@onekeyhq/shared/types/qrCode';
 import type { ETranslations } from '@onekeyhq/shared/src/locale';
 import { openUrlInApp } from '@onekeyhq/shared/src/utils/openUrlUtils';
